Add more escape keys
I want to try out vim and the current esc is a bit far to reach for something I'll be using constantly. So add a couple of esc keys on the left hand with the m01fn key held. I don't know which will be the best, so try out a few positions. m01fn-esc with the left hand can be comfortably followed by shift-: with the right, so that should be good for entering commands.